Introducing a new variable
Introduced a new variable  which allows external images
with SVG extension to be included from whitelisted sites. Default value for
this variable is set to false in DefaultSettings.php owing to security reasons

Wiki owners can set this variable to true in LocalSettings.php to allow SVG
images to be included if they don't care about security.

Added to RELEASE-NOTES-1.25

Bug:T65806